+++ tests/support/MockerExtension.php
@@ -52,6 +52,8 @@ final class MockerExtension implements Extension
             ['namespace' => 'yii\debug\models\router', 'name' => 'is_string'],
             ['namespace' => 'yii\debug\models\router', 'name' => 'count'],
             ['namespace' => 'yii\debug\actions', 'name' => 'sleep'],
+            ['namespace' => 'yii\debug\collectors', 'name' => 'chmod'],
+            ['namespace' => 'yii\debug\collectors', 'name' => 'file_put_contents'],
             ['namespace' => 'PHPForge\Debug\Storage', 'name' => 'file_put_contents'],
             ['namespace' => 'PHPForge\Debug\Storage', 'name' => 'fopen'],
             ['namespace' => 'PHPForge\Debug\Storage', 'name' => 'rename'],